# doc/a3s-openapi.yaml
openapi: 3.0.0
info:
  description: >-
    API Definition for A3S. This service allows authentication, authorisation and accounting.
  version: 1.1.3
  title: A3S
paths:
...
  '/termsOfService':
    get:
      tags:
        - TermsOfService
      description: Search for terms of service entries.
      summary: Search for terms of service entries.
      operationId: listTermsOfServices
      parameters:
        - name: page
          in: query
          required: false
          description: The page to view.
          schema:
            type: integer
            minimum: 1
            format: int32
            default: 1
        - name: size
          in: query
          required: false
          description: The size of a page.
          schema:
            type: integer
            minimum: 1
            maximum: 20
            format: int32
            default: 10
        - name: includeRelations
          in: query
          required: false
          description: Determines whether the related entities, such as teams and which users accepted the agreement, are returned.
          schema:
            type: boolean
            default: false
        - name: filterAgreementName
          in: query
          required: false
          description: A search query filter on the agreement's name.
          schema:
            type: string
            minLength: 0
            maxLength: 255
        - name: orderBy
          in: query
          required: false
          style: form
          explode: false
          description: >-
            a comma separated list of fields in their sort order. Ascending
            order is assumed. Append '_desc' after a field to indicate descending order. Supported fields. 'agreementName'.
          schema:
            type: string
      responses:
        '200':
          description: OK
          headers:
            X-Pagination:
              description: Pagination header.
              schema:
                $ref: '#/components/schemas/PaginationHeaderResponse'
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TermsOfServiceListItem'
        '400':
          description: Bad Request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Not authenticated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden - You are not authorized to access the list of terms of service entries.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: An unexpected error occurred.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      tags:
        - TermsOfService
      summary: Create a new terms of service entry.
      description: Create a new terms of service entry.
      operationId: createTermsOfService
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TermsOfServiceSubmit'
      responses:
        '200':
          description: OK.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TermsOfService'
        '400':
          description: Bad Request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Not authenticated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden - You are not authorized to create terms of service entries.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: An unexpected error occurred.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  '/termsOfService/{termsOfServiceId}':
    get:
      tags:
        - TermsOfService
      summary: Get a terms of service entry.
      description: Get a terms of service entry by its UUID.
      operationId: getTermsOfService
      parameters:
        - name: termsOfServiceId
          in: path
          description: Terms of service entry
          required: true
          schema:
            $ref: '#/components/schemas/TermsOfServiceId'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TermsOfService'
        '400':
          description: Bad Request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Not authenticated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden - You are not authorized to access the terms of service entry.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Terms of service entry not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: An unexpected error occurred
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      tags:
        - TermsOfService
      summary: Delete a terms of service entry.
      description: Deletes a terms of service entry from A3S.
      operationId: deleteTermsOfService
      parameters:
        - name: termsOfServiceId
          description: The UUID of the terms of service entry to delete.
          required: true
          in: path
          schema:
            $ref: '#/components/schemas/TermsOfServiceId'
      responses:
        '204':
          description: No Content.
        '400':
          description: Invalid parameters.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Not authenticated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden - Not authorized to delete terms of service entry.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Terms of service entry not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Terms of service entry cannot be deleted.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: An unexpected error occurred.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  #endregion
  
  # Region - Functions
...
components:
  schemas:
...
    TermsOfServiceId:
      type: string
      format: uuid
    TermsOfService:
      description: |
        Models a terms of service agreement that needs to be agreed to by a user on login.
      type: object
      properties:
        uuid:
          $ref: '#/components/schemas/TermsOfServiceId'
        agreementName:
          type: string
        version:
          type: string
        agreementFileData:
          type: string
        teamIds:
          description: Teams linked to this terms of service entry.
          type: array
          items:
            $ref: '#/components/schemas/TeamId'
        acceptedUserIds:
          description: A list of User Id's that have accepted this terms of service entry.
          type: array
          items:
            $ref: '#/components/schemas/UserId'
        subRealmId:
          $ref: '#/components/schemas/SubRealmId'
    TermsOfServiceListItem:
      description: |
        Models a terms of service agreement returned inside lists.
      type: object
      properties:
        uuid:
          $ref: '#/components/schemas/TermsOfServiceId'
        agreementName:
          type: string
        version:
          type: string
        teamIds:
          description: Teams linked to this terms of service entry.
          type: array
          items:
            $ref: '#/components/schemas/TeamId'
        acceptedUserIds:
          description: A list of User Id's that have accepted this terms of service entry.
          type: array
          items:
            $ref: '#/components/schemas/UserId'
        subRealmId:
          $ref: '#/components/schemas/SubRealmId'
    TermsOfServiceSubmit:
      description: |
        Models a terms of service agreement that needs to be agreed to by a user on login. Used to create or update a terms of service agreement. UUID is required when updating a terms of service agreement. Can be optionally associated with a sub-realm by specifying the 'subRealmId' of the sub-realm.
      required:
        - agreementName
        - agreementFileData
        - autoUpdate
      type: object
      properties:
        uuid:
          $ref: '#/components/schemas/TermsOfServiceId'
        agreementName:
          type: string
        agreementFileData:
          type: string
        autoUpdate:
          type: boolean
        subRealmId:
          $ref: '#/components/schemas/SubRealmId'
      
    # End Region - TermsOfService      

    # Region - Functions
...
    UserId:
      type: string
      format: uuid
...
    TeamId:
      type: string
      format: uuid
...
    SubRealmId:
      description: The UUID identifier for a sub-realm.
      type: string
      format: uuid
...
    ErrorResponse:
      description: Error response object to give more info with HTTP status code
      type: object
      required:
        - message
      properties:
        message:
          type: string
    PaginationHeaderResponse:
      description: >-
        This is the standard response for requests that have paginated
        collections
      type: object
      properties:
        total:
          type: integer
          format: int32
          minimum: 1
          description: The total number of pages in the result set
        count:
          type: integer
          format: int32
          minimum: 1
          description: The total number of results in the result set
        size:
          type: integer
          format: int32
          minimum: 1
          maximum: 20
          default: 10
          description: >-
            The number items to include in a page of results. The page size for
            the result set
        current:
          type: integer
          format: int32
          minimum: 1
          description: >-
            The position of the page in the paged result set that is being
            returned
        first:
          type: string
          format: url
          description: >-
            The link to the first page of results containing [size] results.
            This link includes the page, size, filter, orderBy and fields query
            parameters.
        last:
          type: string
          format: url
          description: >-
            The link to the last page of results containing [size] results. This
            link includes the page, size, filter, orderBy and fields query
            parameters.
        prev:
          type: string
          format: url
          description: >-
            The link to the previous page of results containing [size] results.
            This link includes the page, size, filter, orderBy and fields query
            parameters. If this is the first page in the result set then this
            will be [null].
        next:
          type: string
          format: url
          description: >-
            The link to the next page of results containing [size] results. This
            link includes the page, size, filter, orderBy and fields query
            parameters. If this is the last page in the result set then this will
            be [null].
...